requirements.txt
# Core LangChain and LangGraph dependencies
langgraph>=0.0.50
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0

# Additional dependencies
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Optional: for alternative LLM providers
# langchain-anthropic>=0.1.0
# langchain-google-genai>=0.1.0
